requirements.txt
duckdb==1.2.2
fastapi>=0.104.0
uvicorn>=0.23.2
pyyaml>=6.0
pyarrow>=14.0.0
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0
